refactor: 배치 Step에 날짜 유틸 클래스 적용 - #76
